backend/requirements.txt
djangorestframework
django
django-extensions
django-rest-framework
django-oauth-toolkit
djangorestframework-oauth
